move resolvePathname to swingset-vat
