+++ packages/boot/test/upgrading/test-upgrade-contracts.js
@@ -2,7 +2,7 @@
  * @file cribbed from
  *   packages/zoe/test/swingsetTests/upgradeCoveredCall/test-coveredCall-service-upgrade.js
  */
-import { resolvePathname } from '@agoric/internal';
+import { resolvePathname } from '@agoric/swingset-vat/tools/paths.js';
 import { buildVatController } from '@agoric/swingset-vat';
 import { test as anyTest } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
