+++ packages/builders/scripts/inter-protocol/deploy-contracts.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import { makeHelpers } from '@agoric/deploy-script-support';
 import { E } from '@endo/eventual-send';
-import { resolvePathname } from '@agoric/internal';
+import { resolvePathname } from '@agoric/swingset-vat/tools/paths.js';
 import { getCopyMapEntries, makeCopyMap } from '@agoric/store';
 
 // TODO: CLI options to choose contracts
