+++ packages/internal/src/index.js
@@ -4,7 +4,6 @@
 
 export * from './config.js';
 export * from './debug.js';
-export * from './paths.js';
 export * from './utils.js';
 export * from './method-tools.js';
 export * from './typeGuards.js';
